Added CERN-OHL-W v2
This is a good, balanced license for a project like this.

This license requires that people release any improvements they make to the design, even if they combine it with their proprietary design. However, unless strongly reciprocal licenses, they don't have to open source their proprietary design, and unlike permissive licenses, they do have to offer the source and any improvements taken from this project back to the community.